Some renaming
Production -> SyntaxProduction
IProduction -> Production
ModuleNotFound -> ModuleNotFoundException
UndefinedSymbol -> UndefinedSymbolException
