+++ sdf2table.core/src/main/java/org/metaborg/sdf2table/grammar/ModuleNotFoundException.java
@@ -1,12 +1,12 @@
 package org.metaborg.sdf2table.grammar;
 
-public class ModuleNotFound extends Exception{
+public class ModuleNotFoundException extends Exception{
 	private static final long serialVersionUID = 2895844459531185172L;
 	
 	String _module;
 	Module _required_by;
 	
-	public ModuleNotFound(String module, Module required_by){
+	public ModuleNotFoundException(String module, Module required_by){
 		super("Module not found `"+module+"'"+(required_by == null ? "" : " required by `"+required_by.name()+"'"));
 		_module = module;
 		_required_by = required_by;
