+++ sdf2table.core/src/main/java/org/metaborg/sdf2table/parsetable/Context.java
@@ -6,7 +6,7 @@ import java.util.TreeSet;
 
 import org.metaborg.sdf2table.core.Utilities;
 import org.metaborg.sdf2table.grammar.PriorityLevel;
-import org.metaborg.sdf2table.grammar.Production;
+import org.metaborg.sdf2table.grammar.SyntaxProduction;
 
 public class Context{
 	Set<PriorityLevel> _priorities;
@@ -24,7 +24,7 @@ public class Context{
 		_priorities = new TreeSet<>(other._priorities);
 	}
 	
-	public boolean conflictsLeft(Production p){
+	public boolean conflictsLeft(SyntaxProduction p){
 		if(p.left() != null && p.left().nonEpsilon())
 			return false;
 		for(PriorityLevel l : _priorities){
@@ -34,7 +34,7 @@ public class Context{
 		return false;
 	}
 	
-	public boolean conflictsRight(Production p){
+	public boolean conflictsRight(SyntaxProduction p){
 		if(p.right() != null && p.right().nonEpsilon())
 			return false;
 		for(PriorityLevel l : _priorities){
@@ -44,10 +44,10 @@ public class Context{
 		return false;
 	}
 	
-	public void leftSimplify(Set<Production> derivations){
+	public void leftSimplify(Set<SyntaxProduction> derivations){
 		Set<PriorityLevel> simplified = new TreeSet<>();
 		for(PriorityLevel l : _priorities){
-			for(Production p : derivations){
+			for(SyntaxProduction p : derivations){
 				if(p.left() == null || !p.left().nonEpsilon()){
 					if(l.production().priorities().deepConflicts(p, l.position())){
 						simplified.add(l);
@@ -59,10 +59,10 @@ public class Context{
 		_priorities = simplified;
 	}
 	
-	public void rightSimplify(Set<Production> derivations){
+	public void rightSimplify(Set<SyntaxProduction> derivations){
 		Set<PriorityLevel> simplified = new TreeSet<>();
 		for(PriorityLevel l : _priorities){
-			for(Production p : derivations){
+			for(SyntaxProduction p : derivations){
 				if(p.right() == null || !p.right().nonEpsilon()){
 					if(l.production().priorities().deepConflicts(p, l.position())){
 						simplified.add(l);
